Fix imports invalidated by file renames.
