+++ signac/core/synced_collections/collection_zarr.py
@@ -9,7 +9,7 @@ implementing sync and load methods.
 from copy import deepcopy
 
 from .synced_collection import SyncedCollection
-from .syncedattrdict import SyncedAttrDict
+from .synced_attr_dict import SyncedAttrDict
 from .synced_list import SyncedList
